Refactor plugins module
