+++ src/plugins/mod.rs
@@ -1,14 +1,12 @@
-use ratatui::{backend::CrosstermBackend, layout::Rect, Frame};
-use std::io::Stdout;
-
-pub type PluginFrame<'a> = Frame<'a, CrosstermBackend<Stdout>>;
-
-pub trait PluginRender {
-    fn render(&mut self, f: &mut PluginFrame<'_>, area: Rect);
-}
+pub mod loader;
+pub mod interface;
+pub mod state;
 
 pub mod countdown;
 pub mod pomodoro;
+
+pub use interface::{PluginFrame, PluginRender};
+pub use state::PluginHost;
 
 pub use countdown::CountdownPlugin;
 pub use pomodoro::PomodoroPlugin;
